pyproject.toml
[project]
name = "an"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "alexsuthree", email = "alexsuthree@gmail.com"},
]
dependencies = [
    "djangorestframework>=3.15.2",
    "pymysql>=1.1.1",
    "django-filter>=24.2",
    "django-cors-headers>=4.4.0",
    "toml>=0.10.2",
    "fake-useragent>=1.5.1",
    "cryptography>=43.0.0",
    "django-apscheduler>=0.6.2",
    "ipdb>=0.13.13",
    "pypinyin>=0.52.0",
    "pyexecjs>=1.5.1",
    "requests>=2.32.3",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false


[tool.pdm.scripts]
m1 = "python manage.py makemigrations"
m2 = "python manage.py migrate"